ERP: Do not generate ERP keys when domain name is not specified
This commit adds changes to not generate ERP information if the domain
name is not specified in the EAP identity. keyName-NAI needs the realm
part and as such, it is reasonable to require the main EAP configuration
to provide that realm.

Signed-off-by: Jouni Malinen <jouni@qca.qualcomm.com>
Git-commit: 4aa329298da36c3b477571e30b6ace07b2f38302
Git-repo: git://w1.fi/srv/git/hostap.git
CRs-Fixed: 2050346
Change-Id: I88b64ee35db7d5cc9b82bf3d8a58633ae140f7cc