Interworking: Add NULL checking for EAP name in phase2/autheap parameter
Add NULL checking for EAP name. If it is NULL, do not add the phase2
parameter autheap. This should not happen in practice due to earlier
checks for credential matching, but if there is a code path that would
allow this to be set, it is better to skip setting of the invalid value
and allow automatic selection of the Phase 2 parameters.

Change-Id: Ic1cb6a92ea6756b2b705d17c7cbf1383ada78c90
Signed-off-by: Jouni Malinen <jouni@qca.qualcomm.com>
Git-commit: 8d968351a0a77824a4c703e663427cde767bbd48
Git-repo: git://w1.fi/srv/git/hostap.git
CRs-Fixed: 2053104