wpa_supplicant: add option to configure HIDL service name
When supplicant is built with CONFIG_HIDL it will automatically
register itself as an HIDL service when starting up.
When multiple instances of the supplicant are started, they will
register with the same service name so the last instance started
will overwrite any previous service registrations. This makes it
impossible to control multiple supplicant instances via HIDL.
Fix this by adding a command line option to specify a different
HIDL service name for each supplicant instance.

Change-Id: I545586992fc138edbd81e72f6e65b076ed3073d1