Clear pending_listen_freq if the driver fails to start listen
Current implementation in wpa_supplicant waits a bit if the driver
is unable to start offchannel operation for some reason, with the
assumption that p2p_search() will be started from internal timeout.
But this implementation does not clears pending_listen_freq and as
a result the subsequent listen operations as part of p2p_search
does not happen.
This commit ensures to reset this flag - pending_listen_freq.

CRs-Fixed: 2077492
Change-Id: I6292c235c14ab192e571e9789b023248edb49fb7