mesh: Move writing of mesh_rssi_threshold inside CONFIG_MESH
Previously, the code that writes mesh_rssi_threshold to a network block
always executes, but the code that reads it from network block and the
code that initializes it to a default value in a new network block are
inside #ifdef CONFIG_MESH. As a result when writing a config file it
will write mesh_rssi_threshold (since it has a non-default value) and
later fail to read the network block.

Fix this by moving the write code under #ifdef CONFIG_MESH as well.
Note, network blocks which already have mesh_rssi_threshold because of
the bug will still fail to read after the fix.

Change-Id: Idfa631048dbbccebcf4259699d6e4a746a65690f
Signed-off-by: Lior David <qca_liord@qca.qualcomm.com>
Git-commit: 30311339834bb9f426a36ad31b95860af130a4ae
Git-repo: git://w1.fi/srv/git/hostap.git
CRs-Fixed: 2099771